% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_auto_setup.R
\name{canvas_auto_setup}
\alias{canvas_auto_setup}
\title{Set Up Canvas}
\usage{
canvas_auto_setup()
}
\value{
Sets .Renviron variables and runs \code{\link[=canvas_setup]{canvas_setup()}} to communicate with Canvas
}
\description{
Edits the user's .Renviron file with new or updated environ variables and sets these for use with Canvas.
}
\details{
This interactive function guides the user through the creation (or updating, if they already exist) of two variables, CANVAS_TOKEN and CANVAS_DOMAIN. The function first checks if there are already variables defined in the local .Renviron file with these names, and gives the option to update them or not.
Once both values are put in, the function then sets these for use with Canvas using \code{\link[rcanvas:apihelpers]{rcanvas::set_canvas_token()}} and \code{\link[rcanvas:apihelpers]{rcanvas::set_canvas_domain()}} respectively.
}
\examples{
 \dontrun{
 canvas_auto_setup()
}

}
