% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_manual_setup.R
\name{canvas_manual_setup}
\alias{canvas_manual_setup}
\title{Set Up Canvas Manually}
\usage{
canvas_manual_setup()
}
\value{
Sets .Renviron variables and runs \code{\link[=canvas_setup]{canvas_setup()}} to communicate with Canvas
}
\description{
Edits the user's .Renviron file with new or updated environ variables and sets these for use with Canvas.
}
\examples{
 \dontrun{
 canvas_manual_setup()
}

}
