% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_setup.R
\name{canvas_setup}
\alias{canvas_setup}
\title{Set up access to Canvas}
\usage{
canvas_setup(domain = "https://canvas.sussex.ac.uk")
}
\arguments{
\item{domain}{Canvas url (string)}
}
\value{
Grants canvas access
}
\description{
Set up access to Canvas
}
\examples{
 \dontrun{
 canvas_setup()
}

}
