% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_canvas_data.R
\name{compile_canvas_data}
\alias{compile_canvas_data}
\title{Compile a dataset from Canvas}
\usage{
compile_canvas_data(course_id, sub_id, sub_type = "assignments")
}
\arguments{
\item{course_id}{The Canvas ID number for the course/module}

\item{sub_id}{The Canvas ID number for the assignment}

\item{sub_type}{Type of submission. String. Possible values are "assignments" or "quizzes".}
}
\value{
A tibble of Canvas data combining student infomation, submission information, and sections
}
\description{
Compile a dataset from Canvas
}
\examples{
 \dontrun{
 compile_canvas_data(course_id = 12345, sub_id = 67890)
}

}
