% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sections.R
\name{get_sections}
\alias{get_sections}
\title{Get Canvas sections}
\usage{
get_sections(
  course_id,
  section_regex = "^marking",
  section_number_regex = "^marking\\\\s*(\\\\d+).*"
)
}
\arguments{
\item{course_id}{The Canvas ID number for the course/module}

\item{section_regex}{Regex match as a string for the shared element of the section names that indicate marking groups}

\item{section_number_regex}{Regex string to extract the group number. For example, "^marking\\s*(\\d+).*" identifies the number "1" in "Marking 1 : Take Away paper T2 Week 7"}
}
\value{
A tibble of section information from Canvas
}
\description{
Get Canvas sections
}
\examples{
 \dontrun{
 get_sections(course_id = 12345)
}

}
