% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students.R
\name{get_students}
\alias{get_students}
\title{Get a list of students on a module/course}
\usage{
get_students(course_id)
}
\arguments{
\item{course_id}{The Canvas ID number for the course/module}
}
\value{
A tibble of student information from Canvas
}
\description{
Get a list of students on a module/course
}
\examples{
 \dontrun{
 get_students(course_id = 12345)
}

}
