% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Get submissions for an assignment}
\usage{
get_submissions(course_id, type, type_id)
}
\arguments{
\item{course_id}{Canvas ID number for the course (unit).}

\item{type}{Typo of submission. String. Possible values are "assignments" or "quizzes".}

\item{type_id}{Canvas ID number for the submission. String or numeric.}
}
\value{
a tibble
}
\description{
Get submissions for an assignment
}
\examples{
 \dontrun{
 get_submissions()
}

}
