% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_convenr.R
\name{run_convenr}
\alias{run_convenr}
\title{Run the ConvenR app}
\usage{
run_convenr(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
Runs a Shiny app
}
\description{
Run the ConvenR app
}
\examples{
 \dontrun{
 run_convenr()
}

}
