% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_setup.R
\name{canvas_setup}
\alias{canvas_setup}
\title{Set up access to Canvas}
\usage{
canvas_setup()
}
\value{
Grants canvas access
}
\description{
Set up access to Canvas
}
\examples{
 \dontrun{
 canvas_setup()
}

}
